+++ microfta/test/Data/Tree/FTASpec.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE TypeApplications #-}
 
 module Data.Tree.FTASpec (spec) where
@@ -79,7 +80,7 @@ spec = do
                 shared = Common.intersect naturals evens :: Common.PlainNode String
                 terms = take 9 $ iterate (\term -> Tree.Node "succ" [term]) (Tree.Node "zero" [])
             map (acceptPlain shared) terms `shouldBe` map even [0 :: Int .. 8]
-            let recursiveNodes = Common.crush (\node -> case node of Common.InternedMu _ -> Sum (1 :: Int); _ -> Sum 0)
+            let recursiveNodes = Common.crush (\case Common.InternedMu _ -> Sum (1 :: Int); _ -> Sum 0)
             getSum (recursiveNodes $ Common.Node [Common.Edge "pair" [naturals, naturals]]) `shouldBe` 1
             case Common.toFTA shared of
                 Left err -> expectationFailure $ show err